Finish rich entity domain

+++ src/Services/Ordering/Ordering.Domain/Models/OrderItem.cs
@@ -4,7 +4,7 @@ namespace Ordering.Domain.Models;
 
 public class OrderItem : Entity<OrderItemId>
 {
-    internal OrderItem(OrderId orderId, ProductId productId, decimal price, ushort quantity)
+    internal OrderItem(OrderId orderId, ProductId productId, decimal price, int quantity)
     {
         Id = OrderItemId.Of(Guid.NewGuid());
         OrderId = orderId;
@@ -16,6 +16,6 @@ public class OrderItem : Entity<OrderItemId>
     public OrderId OrderId { get; private set; }
     public ProductId ProductId { get; private set; }
     public decimal Price { get; private set; }
-    public ushort Quantity { get; private set; }
+    public int Quantity { get; private set; }
 
 }
